Rewrite of form driver to allow full dynamic resizing of window
Continued move toward object based design of displays, with data
storage and the previously added size structs added to unified
struct for the form configuration. Multiple pages can be allocated
in the new configuration, but there is no interface to access
additional form pages.

The polling and parsing of data was removed from the form driver
and moved to a separate source file. This will be convenient if
other display options (GTK, command line, etc.) are added later.

The previous strategy to move overflow fields to another page was
removed and instead data storage was moved from the form fields to
a separate data structure. This makes it possible for only the
visible fields to be rendered, then as new fields are rendered they
can get data from the separate storage.