Fixed SIGABRT causing program to exit without NCURSES shutting down
Previously, the program would never exit from the ncurses loop.
Insead all exits were handled by the signal handler. This is no
longer the case and it is expected that on signal the program
will eventually return to the loop for graceful shutdown of ncurses.

This is not the desired or implemented behavior for error signals
like SIGABRT. A emergency exit function was added so that the signal
handler can directly end NCURSES.

Some of the exiting logic in the signal handler was moved to the end
of the main function, as that is where the program should shut down
in normal operation.