Fixed last update not showing on window resize
Since last updated was printed directly to the window, it could
not be restored when re-rendering the window on resize.

This is now stored in a char buffer, and the update routine works
only with this buffer. The refresh routine will automatically
write this to window on any window refresh.